poudriere-image: destroy correct zpool on cleanup
image_zfs uses `zpool create -t` to create an in-core temporary pool
name while the on-disk name will be set to $ZFS_POOL_NAME. By default,
$ZFS_POOL_NAME is set to zroot, to match the zpool name of a standard
FreeBSD install.

If an error occurs while running image_zfs, the on-disk name is used
when destroying the zpool on cleanup, causing the wrong pool to be
destroyed.

Fix this by using the temporary pool name when destroying the zpool on
cleanup.

Fixes:  73b3a516: image_zfs: Allow same-as-host pool names